requirements.txt
discord.py==1.5.1
PyNaCl==1.4.0
